+++ jgiven-core/src/main/java/com/tngtech/jgiven/report/asciidoc/AsciiDocReportConfig.java
@@ -1,13 +1,15 @@
 package com.tngtech.jgiven.report.asciidoc;
 
 import com.tngtech.jgiven.report.AbstractReportConfig;
+import com.tngtech.jgiven.report.config.ConfigOption;
 
+import java.util.List;
 import java.util.Map;
 
 public class AsciiDocReportConfig extends AbstractReportConfig {
 
-    public AsciiDocReportConfig( Map<String, Object> configMap ) {
-        super( configMap );
+    public AsciiDocReportConfig( String... args ) {
+        super( args );
     }
 
     public AsciiDocReportConfig() {
@@ -15,6 +17,10 @@ public class AsciiDocReportConfig extends AbstractReportConfig {
     }
 
     public void useConfigMap( Map<String, Object> configMap ) {
+
+    }
+
+    public void additionalConfigOptions( List<ConfigOption> configOptions ) {
 
     }